% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_factset_db.R
\name{connect_factset_db}
\alias{connect_factset_db}
\title{Export files for use in PACTA data preparation}
\usage{
connect_factset_db(
  dbname = Sys.getenv("PGDATABASE"),
  host = Sys.getenv("PGHOST"),
  port = Sys.getenv("PGPORT", 5432L),
  options = "-c search_path=fds",
  username = Sys.getenv("PGUSER"),
  password = Sys.getenv("PGPASSWORD")
)
}
\arguments{
\item{dbname}{name of the database to connect to}

\item{host}{hostname of the server to connect to}

\item{port}{port number of the server to connect to}

\item{options}{additional options to pass to the database connection.
Typically used to define schema search path.}

\item{username}{username to use for the database connection}

\item{password}{password to use for the database connection}
}
\value{
a database connection object
}
\description{
Export files for use in PACTA data preparation
}
