% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_manifest.R
\name{export_manifest}
\alias{export_manifest}
\title{Export manifest file with metadata}
\usage{
export_manifest(
  manifest_path,
  filelist,
  data_timestamp,
  start_time,
  export_dir
)
}
\arguments{
\item{manifest_path}{Path to the manifest file}

\item{filelist}{List of files to include in the manifest}

\item{data_timestamp}{Timestamp of the data}

\item{start_time}{Timestamp of the start of the export}

\item{export_dir}{Directory to which data was exported}
}
\value{
(invisible) JSON string with metadata manifest.
}
\description{
Export manifest file with metadata
}
