% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_ids.R
\name{get_entity_ids}
\alias{get_entity_ids}
\title{Get the entity info data from the FactSet database and prepare the
\code{factset_entity_info} tibble}
\usage{
get_entity_ids(conn, colname = "factset_entity_id")
}
\arguments{
\item{conn}{database connection}

\item{colname}{column name to pull values from across all tables in database.}
}
\value{
A tibble properly prepared to be saved as the
\code{factset_entity_ids.csv} output file
}
\description{
Get the entity info data from the FactSet database and prepare the
\code{factset_entity_info} tibble
}
