% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factset_isin_to_fund_table.R
\name{get_factset_isin_to_fund_table}
\alias{get_factset_isin_to_fund_table}
\title{Get the isin_to_fund_table data from the FactSet database and prepare the
\code{factset_isin_to_fund_table} tibble}
\usage{
get_factset_isin_to_fund_table(conn)
}
\arguments{
\item{conn}{database connection}
}
\value{
A tibble properly prepared to be saved as the
\code{factset_isin_to_fund_table.rds} output file
}
\description{
Get the isin_to_fund_table data from the FactSet database and prepare the
\code{factset_isin_to_fund_table} tibble
}
