% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fund_data.R
\name{get_fund_data}
\alias{get_fund_data}
\title{Get the fund data from the FactSet database and prepare the
\code{factset_fund_data} tibble}
\usage{
get_fund_data(conn, data_timestamp, data_timestamp_lookback)
}
\arguments{
\item{conn}{databse connection}

\item{data_timestamp}{A date object specifying the desired date for the most
recent data to be included in the output.}

\item{data_timestamp_lookback}{A date object specifying the oldest data
that should be included in the output}
}
\value{
A tibble properly prepared to be saved as the \code{factset_fund_data.rds}
output file
}
\description{
Get the fund data from the FactSet database and prepare the
\code{factset_fund_data} tibble
}
