% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_industry_map_bridge.R
\name{get_industry_map_bridge}
\alias{get_industry_map_bridge}
\title{Get the factset insustry map bridge}
\usage{
get_industry_map_bridge(conn)
}
\arguments{
\item{conn}{database connection}
}
\value{
A tibble properly prepared to be saved as the
\code{factset_industry_map_bridge.rds} output file
}
\description{
Get the factset insustry map bridge
}
