% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iss_emissions_data.R
\name{get_iss_emissions_data}
\alias{get_iss_emissions_data}
\title{Get the ISS emissions data from the FactSet database and prepare the
\code{factset_iss_emissions} tibble}
\usage{
get_iss_emissions_data(
  conn,
  reporting_year,
  min_estimated_trust = 0,
  min_reported_trust = 0
)
}
\arguments{
\item{conn}{databse connection}

\item{min_estimated_trust}{A single numeric specifying the minimum allowed
"estimated trust" value}

\item{min_reported_trust}{A single numeric specifying the minimum allowed
"reported trust" value}

\item{year}{A single numeric specifying the year of data to be returned}

\item{...}{Arguments to be passed to the \code{connect_conn()} function (for
specifying database connection parameters)}
}
\value{
A tibble properly prepared to be saved as the
\code{factset_iss_emissions.rds} output file
}
\description{
Get the ISS emissions data from the FactSet database and prepare the
\code{factset_iss_emissions} tibble
}
