% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_code_bridge.R
\name{get_issue_code_bridge}
\alias{get_issue_code_bridge}
\title{Get the factset issue code bridge}
\usage{
get_issue_code_bridge(conn)
}
\arguments{
\item{conn}{database connection}
}
\value{
A tibble properly prepared to be saved as the
\code{factset_issue_code_bridge.rds} output file
}
\description{
Get the factset issue code bridge
}
