% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.factset-package.R
\docType{package}
\name{workflow.factset-package}
\alias{workflow.factset}
\alias{workflow.factset-package}
\title{workflow.factset: Extract Financial Data for use in PACTA}
\description{
Extract data from a FactSet Postgres database for use as part of PACTA Data Preparation.
}
\author{
\strong{Maintainer}: CJ Yetman \email{cj@cjyetman.com} (\href{https://orcid.org/0000-0001-5099-9500}{ORCID}) [contractor]

Authors:
\itemize{
  \item Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID}) [contractor]
  \item Jacob Kastl \email{jacob.kastl@gmail.com} [contractor]
  \item Alex Axthelm \email{aaxthelm@rmi.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]
}

Other contributors:
\itemize{
  \item RMI \email{PACTA4investors@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
