% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_pacta_files.R
\name{export_pacta_files}
\alias{export_pacta_files}
\title{Export files for use in PACTA data preparation}
\usage{
export_pacta_files(
  destination = file.path("."),
  data_timestamp = Sys.time(),
  ...
)
}
\arguments{
\item{...}{Arguments to be passed to the \code{connect_factset_db()} function (for
specifying database connection parameters)}

\item{Destination}{directory for the output files}
}
\description{
Export files for use in PACTA data preparation
}
