% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factset_entity_info.R
\name{get_factset_entity_info}
\alias{get_factset_entity_info}
\title{Get the entity info data from the FactSet database and prepare the
\code{factset_entity_info} tibble}
\usage{
get_factset_entity_info(...)
}
\arguments{
\item{...}{Arguments to be passed to the \code{connect_factset_db()} function (for
specifying database connection parameters)}
}
\value{
A tibble properly prepared to be saved as the
\code{factset_entity_info.rds} output file
}
\description{
Get the entity info data from the FactSet database and prepare the
\code{factset_entity_info} tibble
}
