% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_manual_sector_override.R
\name{get_manual_sector_override}
\alias{get_manual_sector_override}
\title{Get the PACTA manual sector override table}
\usage{
get_manual_sector_override(conn)
}
\arguments{
\item{conn}{database connection}
}
\value{
A tibble properly prepared to be saved as the
\code{factset_manual_pacta_sector_override.rds} output file
}
\description{
Get the PACTA manual sector override table
}
